requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-cors==5.0.1
flask-limiter[redis]==3.12
psycopg2-binary==2.9.10
redis==5.2.1
python-dotenv==1.1.0
gunicorn==23.0.0
